muchos cambios

+++ app/src/main/java/com/example/athlo/modelo/entidades/EjercicioAsignadoEntity.kt
@@ -9,7 +9,7 @@ data class EjercicioAsignadoEntity(
     val nombre: String = "",
     val series: Int = 0,
     val repeticiones: Int = 0,
-    val peso: Int = 0,
+    val peso: Float = 0f,
     val foto: String = "",
     val video: String = "",
     val idEjercicioFirestore: String = ""
