+++ app/src/main/java/com/example/athlo/modelo/entreno/EjercicioAsignadoRemote.kt
@@ -15,7 +15,7 @@ data class EjercicioAsignadoRemote(
     val nombre: String = "",
     val series: Int = 0,
     val repeticiones: Int = 0,
-    val peso: Int = 0,
+    val peso: Float = 0f,
     val foto: String = "",
     val video: String = ""
 )
